refactor(instrumentation-grpc): clean up remnants of 'grpc' package instrumentation
