+++ experimental/packages/opentelemetry-instrumentation-grpc/src/index.ts
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 
-export * from './instrumentation';
+export { GrpcInstrumentation } from './instrumentation';
 export type { GrpcInstrumentationConfig } from './types';
